Move auth_handler.py to _internal
