--- docs/api/auth_handler.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.auth_handler`
--------------------------------
-
-.. automodule:: certbot.auth_handler
-   :members:
